# TFLM: Added VexRISCV optimized DepthwiseConv2d kernel (int8/uint8)