@advaitjain 2020-08-31
Might be worth linking doc/README.md from the top-level readme.